修改更新

+++ GameFramework/Resource/CheckResourcesCompleteCallback.cs
@@ -1,6 +1,6 @@
 ﻿//------------------------------------------------------------
-// Game Framework v3.x
-// Copyright © 2013-2018 Jiang Yin. All rights reserved.
+// Game Framework
+// Copyright © 2013-2019 Jiang Yin. All rights reserved.
 // Homepage: http://gameframework.cn/
 // Feedback: mailto:jiangyin@gameframework.cn
 //------------------------------------------------------------
@@ -15,5 +15,5 @@ namespace GameFramework.Resource
     /// <param name="updateCount">要更新的资源数量。</param>
     /// <param name="updateTotalLength">要更新的资源总大小。</param>
     /// <param name="updateTotalZipLength">要更新的压缩包总大小。</param>
-    public delegate void CheckResourcesCompleteCallback(bool needUpdateResources, int removedCount, int updateCount, int updateTotalLength, int updateTotalZipLength);
+    public delegate void CheckResourcesCompleteCallback(bool needUpdateResources, int removedCount, int updateCount, long updateTotalLength, long updateTotalZipLength);
 }
